pkg/kube/forwarder_test.go
package kube

import (
	"github.com/stretchr/testify/assert"
	"testing"
)

// Mocked forwarder structure for demonstration. Replace it with your actual forwarder implementation.
type Forwarder struct {
	// Fields of the forwarder
}

// Mocked method for demonstration. Replace it with your actual Forward method.
func (f *Forwarder) Forward(data string) string {
	// Placeholder logic
	return data
}

func TestForwarder(t *testing.T) {
	// Creating an instance of forwarder, replace it as per your structure
	f := &Forwarder{}

	// Test cases
	tests := []struct {
		name     string
		input    string
		expected string
	}{
		{"TestEmptyInput", "", ""},
		{"TestStandardInput", "hello", "hello"},
		{"TestSpecialCharacters", "@!#$", "@!#$"},
	}

	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			output := f.Forward(tt.input)
			assert.Equal(t, tt.expected, output, "they should be equal")
		})
	}
}
